# Trivial parser translator optimizations