I ran it over the RuboCop codebase. Here are some numbers:

Before:
```
 hyperfine -w 2 -r 10 "bundle exec ruby test.rb"
Benchmark 1: bundle exec ruby test.rb
  Time (mean ± σ):     24.055 s ±  0.410 s    [User: 23.661 s, System: 0.305 s]
  Range (min … max):   23.464 s … 24.610 s    10 runs
```

After:
```
$ hyperfine -w 2 -r 10 "bundle exec ruby test.rb"
Benchmark 1: bundle exec ruby test.rb
  Time (mean ± σ):     21.464 s ±  0.447 s    [User: 21.156 s, System: 0.220 s]
  Range (min … max):   20.888 s … 22.240 s    10 runs
```

So, about 11% faster just with this. The majority of the time is still spent translating, only ~10% is actually spend by prism parsing/lexing it.

Benchmark file:
```rb
require "prism"
require "prism/translation/parser"


Dir.glob("../rubocop/**/*.rb").each do
  buffer = Parser::Source::Buffer.new(it, 1)
  buffer.source = File.read(it)
  Prism::Translation::Parser33.new.tokenize(buffer)
end
```