@Earlopain 2025-03-11
Not quite sure what the failures are about. I guess it doesn't like `TESTOPTS` being specified anymore?