@Earlopain 2025-03-11
Bisected to https://github.com/ruby/optparse/commit/050a87d02915400ee67ca6650dbaa131732f7c73. I commented on the PR